backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
chromadb==1.5.5
pypdf==6.9.2
google-genai==1.69.0
python-dotenv==1.2.2
python-multipart==0.0.22
pydantic==2.12.5
